refactor: restructure project for better integration of new services
