+++ main.go
@@ -4,7 +4,15 @@ import (
 	"log"
 	"os"
 
-	"github.com/kha7iq/pingme/cmd"
+	"github.com/kha7iq/pingme/service/discord"
+	"github.com/kha7iq/pingme/service/email"
+	"github.com/kha7iq/pingme/service/mattermost"
+	"github.com/kha7iq/pingme/service/msteams"
+	"github.com/kha7iq/pingme/service/pushbullet"
+	"github.com/kha7iq/pingme/service/pushover"
+	"github.com/kha7iq/pingme/service/rocketchat"
+	"github.com/kha7iq/pingme/service/slack"
+	"github.com/kha7iq/pingme/service/telegram"
 
 	"github.com/urfave/cli/v2"
 )
@@ -12,7 +20,7 @@ import (
 // Version variable is used for semVer
 var Version string
 
-// main with combile all the function into commands
+// main with combine all the function into commands
 func main() {
 	app := cli.NewApp()
 	app.Name = "PingMe"
@@ -24,15 +32,15 @@ variables and command line switches.Currently supported platforms include Slack,
 RocketChat, Discord, Pushover, Mattermost, Pushbullet, Microsoft Teams and email address.`
 	// app.Commands contains the subcommands as functions which return []*cli.Command.
 	app.Commands = []*cli.Command{
-		cmd.SendToTelegram(),
-		cmd.SendToRocketChat(),
-		cmd.SendToSlack(),
-		cmd.SendToDiscord(),
-		cmd.SendToTeams(),
-		cmd.SendToPushOver(),
-		cmd.SendToEmail(),
-		cmd.SendToMattermost(),
-		cmd.SendToPushBullet(),
+		telegram.Send(),
+		rocketchat.Send(),
+		slack.Send(),
+		discord.Send(),
+		msteams.Send(),
+		pushover.Send(),
+		email.Send(),
+		mattermost.Send(),
+		pushbullet.Send(),
 	}
 
 	err := app.Run(os.Args)
