+++ service/slack/slack.go
@@ -1,10 +1,11 @@
-package cmd
+package slack
 
 import (
 	"context"
-	"fmt"
 	"log"
 	"strings"
+
+	"github.com/kha7iq/pingme/service/helpers"
 
 	"github.com/nikoksr/notify"
 	"github.com/nikoksr/notify/service/slack"
@@ -19,11 +20,11 @@ type slackPingMe struct {
 	Title   string
 }
 
-// SendToSlack parse values from *cli.context and return *cli.Command.
+// Send parse values from *cli.context and return *cli.Command.
 // Values include slack token, channelIDs, Message and Title.
 // If multiple channels are provided then the string is split with "," separator and
 // each channelID is added to receiver.
-func SendToSlack() *cli.Command {
+func Send() *cli.Command {
 	var slackOpts slackPingMe
 	return &cli.Command{
 		Name:      "slack",
@@ -59,7 +60,7 @@ All configuration options are also available via environment variables.`,
 			&cli.StringFlag{
 				Destination: &slackOpts.Title,
 				Name:        "title",
-				Value:       TimeValue,
+				Value:       helpers.TimeValue,
 				Usage:       "Title of the message.",
 				EnvVars:     []string{"SLACK_MSG_TITLE"},
 			},
@@ -70,7 +71,7 @@ All configuration options are also available via environment variables.`,
 			chn := strings.Split(slackOpts.Channel, ",")
 			for _, v := range chn {
 				if len(v) <= 0 {
-					return fmt.Errorf(EmptyChannel)
+					return helpers.ErrChannel
 				}
 
 				slackSvc.AddReceivers(v)
